PMM-4192 Updated MongoDB structs (#174)
* PMM-4192 Explain for MongoDB

Added explain for MongoDB to the pmm-agent actions

* PMM-4192 Fixes suggested by linter

* PMM-4192 Fixed mongodb explain returned Type()

* PMM-4192 Improved MongoExplain test

* PMM-4192 Add license headers.

* PMM-4192 Removed PT dependency

Copied RunExplain function to avoid importing an extra PT package

* PMM-4192 Fixed test

* PMM-4192 Updated code according to CR

* PMM-4192 Updated deps

* PMM-4192 Fixex MongoDB explain

The Explain action for MongodB was not being handled by the agent
(missing case in switch)

* PMM-4192 Updated pmm pkg & fixed db

Explain must be executed against the admin db. I removed the Database
field from the request struct and now it is fixed to admin

* PMM-4192 Fixed test

* PMM-4192 Added proper comments to exported structs and methods

* PMM-4192 Moved defer to correct location

* PMM-4192 Remove double spaces.

* PMM-4192 Address requested changes.

* PMM-4192 Fix tests.

* PMM-4192 Fix tests.

* PMM-4192 Removed redundant check

Removed MySQL flavor comparisson for MySQL 8 because there is only one
vendor for v8.0
Also updated the way we get the vendor name from mysql version_comment
since not all vendors and versions have the same capitalizacion in the
name. Now the name is being converted to lowercase before comparing.

* PMM-4192 Updated Mongo system profile structs

* PMM-4192 go-formatted

* PMM-4192 Updated toolkit deps

* Trying to fix explain

* wip research 4 new test

* PMM-4192 Added mongo explain test

* PMM-4192 Removed debug code

* PMM-4192 Updated deps

* PMM-4192 Removed debug code

Co-authored-by: Alexey Palazhchenko <alexey.palazhchenko@percona.com>
Co-authored-by: Nurlan Moldomurov <nurlan.moldomurov@percona.com>
Co-authored-by: Carlos <cfsalguero@gmail.com>
Co-authored-by: Andrii Skomorokhov <askomorokhov@users.noreply.github.com>