Feedback from forum user (#559)
A user on the forums interpreted this as a command they must run on a newly created PMM server from AWS Marketplace and broke their install badly!  Needed to clarify that is an example only for another DB host and should NOT be run on the PMM server.  Feel free to reword!