clarify what is needed to bypass security warning
Chrome users that do not use "localhost" as the address to test from will not see an option to proceed from the security warning and must type a secret code to get past it.  this is NOT intuitive as there's no input box to type into...trying to make that clearer in the output.