PMM-9413 Fix deadlock. (#982)
In PMM server, locks are used to update list of agents. Agent addition or removal is performed after lock is acquired. Due to programming error, we hit a scenario where we perform removal of an agent while adding another agent. This leads to deadlock, as goroutine waits to acquire a lock that it already holds.

This change fixes the bug. We now acquire lock for agent addition only after agent removal is performed.