replace timerange implementation with pbm status
the current way of listing pitr timeranges for MongoDB seem to have some
limitations e.g
- we can't detect timeranges with no base snapshot
- we can't correctly detect replica sets and return all timestamps
additionally, making changes (such as adding support for local storage)
will involve a lot of code duplication from pbm.

This change replaces the current implementation with a call to pbm
status. This fixes the issue above, but with the extra requirement of
having pbm installed on the pmm-agent.