Update mongodb_limitations.md (#923)
* Update mongodb_limitations.md

I find it really difficult to read that massive table with what is supported and what is not. I suggest we remove two columns and do 4 shorter tables instead. That way it's much easier to navigate and find what you need.

* Update mongodb_limitations.md

I noticed that the column for containerized can be removed as well - so simplified it further

---------

Co-authored-by: Rasika Chivate <95711051+rasika-chivate@users.noreply.github.com>