PMM-11975 Temporarily disable dbaas postgresql monitoring
With the release of PG operator v2.1.0 a bug was introduced that
inhibits the creation of the secret that contains the host/password to
access the cluster when PMM monitoring is enabled.
Until a workaround for this is found, we disable the monitoring as it's
less critical than having the credentials to access the cluster.