PMM-12155 Use new option name for node_exporter (#2186)
* PMM-12155 Use new option name for node_exporter

Since the new version of exporter-toolkit used in the
node exporter the new option `--web.config.file` should be used
instead to pass the TLS configuration.

* PMM-12155 Improve formatting

Improve variable declaration formatting.

* PMM-12155 Add unit test for new option

Add unit test to pass the new option to the exporter and
add some slight fixes to the formatting and styling.

* PMM-12155 Styling fixes

* PMM-12155 Styling fixes

* PMM-12155 refactoring.

* PMM-12155 refactoring.

* PMM-12155 fix linter.

* PMM-12155 fix build.

* PMM-12155 fix linter.

* Update version/features.go

---------

Co-authored-by: Roman Novikov <roma.novikov@percona.com>
Co-authored-by: Nurlan Moldomurov <nurlan.moldomurov@percona.com>
Co-authored-by: Nurlan Moldomurov <bupychuk1989@gmail.com>