Changes in docker commands used.
The volume path was incorrect for the pmm-client-data container, as it was using /srv/ when it should be using /usr/local/percona/pmm2/.

Also, what was mentioned about -d was incorrect, and using --rm as it was meant that the container will be removed when the process is stopped. It's better to use -d and leave it running in the background, and docker logs will still work in this case, so we don't lose any functionality.