Changes in docker commands used. (#4046)
* Changes in docker commands used.

The volume path was incorrect for the pmm-client-data container, as it was using /srv/ when it should be using /usr/local/percona/pmm2/.

Also, what was mentioned about -d was incorrect, and using --rm as it was meant that the container will be removed when the process is stopped. It's better to use -d and leave it running in the background, and docker logs will still work in this case, so we don't lose any functionality.

* Update documentation/docs/setting-up/client/index.md

Co-authored-by: Nurlan Moldomurov <nurlan.moldomurov@percona.com>

---------

Co-authored-by: Catalina A <94133018+catalinaadam@users.noreply.github.com>
Co-authored-by: Nurlan Moldomurov <nurlan.moldomurov@percona.com>